ntt/*: make it work with earlier ROCm compiler.

+++ ntt/kernels/ct_mixed_radix_narrow.cu
@@ -93,7 +93,7 @@ void _CT_NTT(fr_t* d_inout, const unsigned int lg_domain_size,
     }
     noop();
 
-    unsigned int rev_idx = bit_rev(threadIdx.x, iterations);
+    unsigned int rev_idx = bit_rev((unsigned int)threadIdx.x, iterations);
 
     #pragma unroll 1
     for (unsigned int s = 1; s < min(iterations, 6u); s++) {
